+++ src/interface/interface.js
@@ -3,6 +3,8 @@ const KEYS = require('./keys')
 const terminalStyle = require('./terminalStyle')
 const EventEmitter = require('events')
 const debug = require('debug')('interface')
+const MainScreen = require('./screens/mainScreen')
+
 
 // TODO: rename to screen and instead of INDEX call it MAIN
 const WINDOWS = {
@@ -25,6 +27,7 @@ class Interface extends EventEmitter {
   numberOfLines = 0
   cursorPosition = 0
   lock = false
+  screen = null
   window = WINDOWS.INDEX
 
   // FIXME
@@ -48,12 +51,9 @@ class Interface extends EventEmitter {
     this.sendTransaction = args.sendTransaction
     this.store = args.store
 
+    this.screen = new MainScreen({store: args.store})
     this._init()
 
-    this.store.on('changed', () => {
-      this.update()
-    })
-    
     this.store.on('rejected', () => {
       if (this.window === WINDOWS.SEND_DOGECOINS) {
         if (this.store.rejectMessage.message === 'tx') {
@@ -70,14 +70,6 @@ class Interface extends EventEmitter {
         case KEYS.CTRL_C:
           this._stop()
           break
-
-        // FIXME: the delay can be annoying so we let this down for now
-        /*case KEYS.UP:
-          this._updateMenuSelection(-1)
-          break
-        case KEYS.DOWN:
-          this._updateMenuSelection(1)
-          break*/
         case KEYS.NUM_KEY_1:
           this.window = WINDOWS.GENERATE_ADDRESS
           this._displayGenerateAddressWindow()
@@ -94,7 +86,7 @@ class Interface extends EventEmitter {
           this.window = WINDOWS.INDEX
           process.stdout.moveCursor(this.cursorPosition, -(this.numberOfLines-1), () => {
             process.stdout.write(terminalStyle.CLEAR)
-            process.stdout.write(this.format(
+            process.stdout.write(this.screen.format(
               this.store.height,
               this.store.bestHeight,
               this.store.hash,
@@ -121,7 +113,7 @@ class Interface extends EventEmitter {
     process.stdout.write(terminalStyle.NO_CURSOR)
     process.stdout.write('\x1b]0;Dogecoin SPV node wallet\x07')
 
-    process.stdout.write(this.format())
+    process.stdout.write(this.screen.format())
 
     // without this, we would only get streams once enter is pressed
     process.stdin.setRawMode(true)
@@ -134,25 +126,20 @@ class Interface extends EventEmitter {
     process.stdin.setEncoding('utf-8')
   }
 
-  _unlock () {
-    this.lock = false
-    this.emit('unlock')
-  }
-
   _stop () {
     // Need to have screen unlock (so no update)
     this.isShuttingDown = true
 
-    if (this.lock) {
-      this.on('unlock', this._quit)
+    if (this.screen.lock) {
+      this.screen.on('unlock', this._quit)
     } else {
-      this.lock = true
+      this.screen.lock = true
       this._quit()
     }
   }
 
   _quit () {
-    process.stdout.moveCursor(this.cursorPosition, -(this.numberOfLines-1), () => {
+    process.stdout.moveCursor(this.screen.cursorPosition, -(this.screen.numberOfLines-1), () => {
       process.stdout.write(`${terminalStyle.CLEAR}${terminalStyle.SHOW_CURSOR}`)
       // clean screen then quit
       this.emit('quit')
@@ -210,89 +197,6 @@ class Interface extends EventEmitter {
 
       process.stdout.write(layout)
     })
-  }
-
-
-  // FIXME: the delay can be annoying so we let this down for now
-  _updateMenuSelection (direction) {
-
-    // get current position
-    let newPosition = (this.menuSelection['currentPosition'] + direction) % 3
-
-    if (newPosition < 0) {
-      newPosition = 2
-    }
-
-    console.log("NEW POSITION : " + newPosition)
-
-    // clean
-    // FIXME
-    this.menuSelection['0'] = ''
-    this.menuSelection['1'] = ''
-    this.menuSelection['2'] = ''
-
-    this.menuSelection['currentPosition'] = newPosition
-    this.menuSelection[newPosition] = MENU_PREFIX
-
-  }
-
-  // Update interface
-  update () {
-    if (this.lock) { return }
-    if (this.window !== WINDOWS.INDEX) { return }
-
-    this.lock = true
-
-    //  TODO: properly get position of each value and only update it instead of the all screen
-    process.stdout.moveCursor(this.cursorPosition, -(this.numberOfLines-1), () => {
-      process.stdout.write(this.format(
-        this.store.height,
-        this.store.bestHeight,
-        this.store.hash,
-        this.store.getNumPeers(),
-        this.store.tips,
-        this.store.merkleHeight,
-        this.store.balance
-      ))
-
-      // Unlock interface
-      this._unlock()
-    })
-  }
-
-  format (height=0, bestHeight=0, hash=null, numberOfPeers=0, tips=new Map(), merkleHeight=0, balance=0) {
-    const report = process.resourceUsage()
-    const rss = Math.floor(process.memoryUsage().rss/(1000*1000))
-
-    // TODO: seperate in sublayout
-
-    const layout = `
-================ Process Usage Report ================
-
-    fsRead: ${report.fsRead}  fsWrite: ${report.fsWrite}
-    Memory usage: ${rss} MB
-
-================ SPV node ============================
-
-    Height headers: ${height}/${bestHeight}
-    Hash: ${hash}
-    Peers: ${numberOfPeers}
-    Tips: ${JSON.stringify([...tips.keys()])}
-    Merkle Height: ${merkleHeight}/${bestHeight}
-
-================ Wallet =============================
-
-    Balance: ${balance/constants.SATOSHIS} Ð
-
-================ Menu ===============================
-
-    1. ${this.menuSelection['0']}Generate a new address${MENU_SUFFIX}
-    2. ${this.menuSelection['1']}Send dogecoins${MENU_SUFFIX}
-    3. ${this.menuSelection['2']}Quit${MENU_SUFFIX}
-`
-    this.numberOfLines = layout.split('\n').length
-
-    return layout
   }
 
 }
