+++ src/interface/screens/sendDogeScreen.js
@@ -1,0 +1,6 @@
+const EventEmitter = require('events')
+
+class SendDogeScreen extends EventEmitter {
+}
+
+module.exports = SendDogeScreen
